% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_first_cluster}
\alias{get_first_cluster}
\title{Get first cluster}
\usage{
get_first_cluster(distance_table, verbose = F)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
In the first iteration all the objects represent individual clusters.
Therefore all the linkage method would decide the same winner.
For this reason is not necessary to create and aggregate the rankings
as the row with the closest distance is the one selected.
}
