% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{init_distance_table}
\alias{init_distance_table}
\title{Init distance table}
\usage{
init_distance_table(data)
}
\arguments{
\item{data}{table with data for computing distance}
}
\value{

}
\description{
The distance matrix is compute for the `data` given as argument using
the `base::dist` function.
}
\details{
The aim of this function is to create a table of the format
cFrom | cTo | distance
that computes in a tidy format for each pair of clusters the distance
between them.
}
\examples{
set.seed(123)
# Artificial data for five objects
obj <- tibble::tibble(x = sample(1:10,5), y = sample(1:10,5))
# Each object is one cluster, pairwise comparison of the clusters given by
init_distance_matrix(obj)

}
