% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mchlust.R
\name{mc_hclust}
\alias{mc_hclust}
\title{Multicriteria hirarchical clustering}
\usage{
mc_hclust(data, linkage_methods, aggregation_method = "borda", verbose = F)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Multicriteria hirarchical clustering
}
