% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_clusters}
\alias{merge_clusters}
\title{Merge clusters}
\usage{
merge_clusters(distance_table, merge1, merge2)
}
\arguments{
\item{distance_table}{tidy distance table as the one given by
`init_distance_matrix`.}

\item{merge1}{The index of the first cluster to merge}

\item{merge2}{The index of the second cluster to merge}
}
\value{
The new distance table for the upper level of the dendrogram
}
\description{
This recieves a distance table and merge the clusters together
}
