% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_viz.R
\name{plot_mchclust_tiles}
\alias{plot_mchclust_tiles}
\title{Plot clustering tiles}
\usage{
plot_mchclust_tiles(mchc, filter = 0)
}
\arguments{
\item{objects}{}
}
\value{

}
\description{
For objects created with the `mc_hclust` function
}
